Uses consistent scale-down algorithm for both packing plans
